+++ great_expectations/expectations/metrics/table_metrics/table_head.py
@@ -4,6 +4,9 @@ from typing import TYPE_CHECKING, Any, Iterator
 
 import pandas as pd
 
+from great_expectations.compatibility.sqlalchemy_compatibility_wrappers import (
+    read_sql_table_as_df,
+)
 from great_expectations.core.metric_domain_types import MetricDomainTypes
 from great_expectations.execution_engine import (
     PandasExecutionEngine,
@@ -16,7 +19,6 @@ from great_expectations.expectations.metrics.table_metric_provider import (
     TableMetricProvider,
 )
 from great_expectations.optional_imports import sqlalchemy as sa
-from great_expectations.sqlalchemy_compatibility_wrappers import read_sql_table_as_df
 from great_expectations.validator.metric_configuration import MetricConfiguration
 from great_expectations.validator.validator import Validator
