+++ tests/data_context/test_data_context_test_yaml_config.py
@@ -15,7 +15,9 @@ from great_expectations.data_context.config_validator.yaml_config_validator impo
 )
 from great_expectations.data_context.store import CheckpointStore
 from great_expectations.data_context.util import file_relative_path
-from great_expectations.sqlalchemy_compatibility_wrappers import add_dataframe_to_db
+from great_expectations.compatibility.sqlalchemy_compatibility_wrappers import (
+    add_dataframe_to_db,
+)
 from great_expectations.rule_based_profiler.rule_based_profiler import RuleBasedProfiler
 from great_expectations.util import get_sqlalchemy_url, load_class
 from tests.core.usage_statistics.util import (
