+++ tests/datasource/fluent/integration/conftest.py
@@ -22,7 +22,9 @@ from great_expectations.datasource.fluent.interfaces import (
 from great_expectations.datasource.fluent.sources import (
     DEFAULT_PANDAS_DATA_ASSET_NAME,
 )
-from great_expectations.sqlalchemy_compatibility_wrappers import add_dataframe_to_db
+from great_expectations.compatibility.sqlalchemy_compatibility_wrappers import (
+    add_dataframe_to_db,
+)
 
 logger = logging.getLogger(__name__)
