+++ tests/execution_engine/split_and_sample/test_sqlalchemy_execution_engine_sampling.py
@@ -19,7 +19,9 @@ from great_expectations.execution_engine.sqlalchemy_batch_data import (
 from great_expectations.execution_engine.sqlalchemy_dialect import GXSqlDialect
 from great_expectations.self_check.util import build_sa_engine
 from great_expectations.util import import_library_module
-from great_expectations.sqlalchemy_compatibility_wrappers import add_dataframe_to_db
+from great_expectations.compatibility.sqlalchemy_compatibility_wrappers import (
+    add_dataframe_to_db,
+)
 
 try:
     sqlalchemy = pytest.importorskip("sqlalchemy")
