+++ tests/profile/test_user_configurable_profiler_v3_batch_request.py
@@ -28,7 +28,9 @@ from great_expectations.self_check.util import (
     connection_manager,
     get_sql_dialect_floating_point_infinity_value,
 )
-from great_expectations.sqlalchemy_compatibility_wrappers import add_dataframe_to_db
+from great_expectations.compatibility.sqlalchemy_compatibility_wrappers import (
+    add_dataframe_to_db,
+)
 from great_expectations.util import is_library_loadable
 from great_expectations.validator.validator import Validator
 from tests.profile.conftest import get_set_of_columns_and_expectations_from_suite
